Open sourcing the Gorgeous Code Project
This is a project I started long time ago (during my master thesis).
Over time the project took complete changes in direction and the current
result can be seen https://gorgeouscode.com/.

As a result of lack of time it was never possible to evolve it further
enough. So now even if the code is a little bit fuzzy I decided to open
source the code base, and have freeletics sponsoring the hosting for it.

I would like to thank Miguel Fernandes, André Santos, Pedro Rangel Henriques for their contributions.